updated user model charge left to floatfield, added get_length func

+++ converter/users/models.py
@@ -1,5 +1,5 @@
 from django.contrib.auth.models import AbstractUser
-from django.db.models import CharField, IntegerField
+from django.db.models import CharField, FloatField
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
@@ -11,7 +11,7 @@ class User(AbstractUser):
     name = CharField(_("Name of User"), blank=True, max_length=255)
     first_name = None  # type: ignore
     last_name = None  # type: ignore
-    convert_min_left = IntegerField(
+    convert_min_left = FloatField(
         _("How many minutes of video user can convert"), default=1000
     )
